refactor(core): move state-fingerprint helpers to block_executor submodule
Pure code motion, no behavior change: emit_apply_profile, emit_state_fingerprint, compute_state_fingerprint, StateComponents and state_components move out of the 3390-line block_executor.rs into block_executor/fingerprint.rs (continuing the existing block_executor/{evm,validate} split). These are debug/observability helpers gated behind SENTRIX_STATE_FINGERPRINT / SENTRIX_APPLY_PROFILE — they never feed consensus. state_components + StateComponents stay pub (re-exported, used by the bin's state command); the rest are pub(crate). Also un-scrambles the doc comments that had drifted above the wrong function. cargo check --workspace -D warnings clean; sentrix-core 273 tests pass (unchanged).
